fix(flashbots-rpc): tests
