+++ crates/flashblocks-rpc/tests/rpc.rs
@@ -12,16 +12,22 @@ use alloy_rpc_client::RpcClient;
 use alloy_rpc_types::simulate::{SimBlock, SimulatePayload};
 use alloy_rpc_types_engine::PayloadId;
 use alloy_rpc_types_eth::{TransactionInput, error::EthRpcErrorCode};
-use base_reth_flashblocks_rpc::subscription::{Flashblock, Metadata};
+use base_reth_flashblocks_rpc::{
+    pubsub::{BasePubSub, BasePubSubApiServer},
+    subscription::{Flashblock, Metadata},
+};
 use base_reth_test_utils::flashblocks_harness::FlashblocksHarness;
 use common::{BLOCK_INFO_TXN, BLOCK_INFO_TXN_HASH};
 use eyre::Result;
+use futures_util::{SinkExt, StreamExt};
 use op_alloy_consensus::OpDepositReceipt;
 use op_alloy_network::{Optimism, ReceiptResponse, TransactionResponse};
 use op_alloy_rpc_types::OpTransactionRequest;
 use reth_optimism_primitives::OpReceipt;
 use reth_rpc_eth_api::RpcReceipt;
 use rollup_boost::{ExecutionPayloadBaseV1, ExecutionPayloadFlashblockDeltaV1};
+use serde_json::json;
+use tokio_tungstenite::{connect_async, tungstenite::Message};
 
 struct TestSetup {
     harness: FlashblocksHarness,
@@ -728,6 +734,182 @@ async fn test_get_logs_mixed_block_ranges() -> Result<()> {
     // Should include pending logs (historical part is empty in our test setup)
     assert_eq!(logs.len(), 2);
     assert!(logs.iter().all(|log| log.transaction_hash == Some(INCREMENT_HASH)));
+
+    Ok(())
+}
+
+// base_ methods
+#[tokio::test]
+async fn test_base_subscribe_new_flashblocks() -> eyre::Result<()> {
+    let setup = TestSetup::new().await?;
+    let provider = setup.harness.provider();
+    let ws_url = setup.harness.ws_url();
+    let (mut ws_stream, _) = connect_async(&ws_url).await?;
+
+    ws_stream
+        .send(Message::Text(
+            json!({
+                "jsonrpc": "2.0",
+                "id": 1,
+                "method": "base_subscribe",
+                "params": ["newFlashblocks"]
+            })
+            .to_string()
+            .into(),
+        ))
+        .await?;
+
+    let response = ws_stream.next().await.unwrap()?;
+    let sub: serde_json::Value = serde_json::from_str(response.to_text()?)?;
+    assert_eq!(sub["jsonrpc"], "2.0");
+    assert_eq!(sub["id"], 1);
+    let subscription_id = sub["result"].as_str().expect("subscription id expected");
+
+    setup.send_flashblock(create_first_payload()).await?;
+
+    let notification = ws_stream.next().await.unwrap()?;
+    let notif: serde_json::Value = serde_json::from_str(notification.to_text()?)?;
+    assert_eq!(notif["method"], "base_subscription");
+    assert_eq!(notif["params"]["subscription"], subscription_id);
+
+    let block = &notif["params"]["result"];
+    assert_eq!(block["number"], "0x1");
+    assert!(block["hash"].is_string());
+    assert!(block["parentHash"].is_string());
+    assert!(block["transactions"].is_array());
+    assert_eq!(block["transactions"].as_array().unwrap().len(), 1);
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn test_base_subscribe_multiple_flashblocks() -> eyre::Result<()> {
+    let setup = TestSetup::new().await?;
+    let provider = setup.harness.provider();
+    let ws_url = setup.harness.ws_url();
+    let (mut ws_stream, _) = connect_async(&ws_url).await?;
+
+    ws_stream
+        .send(Message::Text(
+            json!({
+                "jsonrpc": "2.0",
+                "id": 1,
+                "method": "base_subscribe",
+                "params": ["newFlashblocks"]
+            })
+            .to_string()
+            .into(),
+        ))
+        .await?;
+
+    let response = ws_stream.next().await.unwrap()?;
+    let sub: serde_json::Value = serde_json::from_str(response.to_text()?)?;
+    let subscription_id = sub["result"].as_str().expect("subscription id expected");
+
+    setup.send_flashblock(create_first_payload()).await?;
+
+    let notif1 = ws_stream.next().await.unwrap()?;
+    let notif1: serde_json::Value = serde_json::from_str(notif1.to_text()?)?;
+    assert_eq!(notif1["params"]["subscription"], subscription_id);
+
+    let block1 = &notif1["params"]["result"];
+    assert_eq!(block1["number"], "0x1");
+    assert_eq!(block1["transactions"].as_array().unwrap().len(), 1);
+
+    setup.send_flashblock(create_second_payload()).await?;
+
+    let notif2 = ws_stream.next().await.unwrap()?;
+    let notif2: serde_json::Value = serde_json::from_str(notif2.to_text()?)?;
+    assert_eq!(notif2["params"]["subscription"], subscription_id);
+
+    let block2 = &notif2["params"]["result"];
+    assert_eq!(block1["number"], block2["number"]); // Same block, incremental updates
+    assert_eq!(block2["transactions"].as_array().unwrap().len(), 6);
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn test_base_unsubscribe() -> eyre::Result<()> {
+    let setup = TestSetup::new().await?;
+    let provider = setup.harness.provider();
+    let ws_url = setup.harness.ws_url();
+    let (mut ws_stream, _) = connect_async(&ws_url).await?;
+
+    ws_stream
+        .send(Message::Text(
+            json!({
+                "jsonrpc": "2.0",
+                "id": 1,
+                "method": "base_subscribe",
+                "params": ["newFlashblocks"]
+            })
+            .to_string()
+            .into(),
+        ))
+        .await?;
+
+    let response = ws_stream.next().await.unwrap()?;
+    let sub: serde_json::Value = serde_json::from_str(response.to_text()?)?;
+    let subscription_id = sub["result"].as_str().expect("subscription id expected");
+
+    ws_stream
+        .send(Message::Text(
+            json!({
+                "jsonrpc": "2.0",
+                "id": 2,
+                "method": "base_unsubscribe",
+                "params": [subscription_id]
+            })
+            .to_string()
+            .into(),
+        ))
+        .await?;
+
+    let unsub = ws_stream.next().await.unwrap()?;
+    let unsub: serde_json::Value = serde_json::from_str(unsub.to_text()?)?;
+    assert_eq!(unsub["jsonrpc"], "2.0");
+    assert_eq!(unsub["id"], 2);
+    assert_eq!(unsub["result"], true);
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn test_base_subscribe_multiple_clients() -> eyre::Result<()> {
+    let setup = TestSetup::new().await?;
+    let provider = setup.harness.provider();
+    let ws_url = setup.harness.ws_url();
+    let (mut ws1, _) = connect_async(&ws_url).await?;
+    let (mut ws2, _) = connect_async(&ws_url).await?;
+
+    let req = json!({
+        "jsonrpc": "2.0",
+        "id": 1,
+        "method": "base_subscribe",
+        "params": ["newFlashblocks"]
+    });
+    ws1.send(Message::Text(req.to_string().into())).await?;
+    ws2.send(Message::Text(req.to_string().into())).await?;
+
+    let _sub1 = ws1.next().await.unwrap()?;
+    let _sub2 = ws2.next().await.unwrap()?;
+
+    setup.send_flashblock(create_first_payload()).await?;
+
+    let notif1 = ws1.next().await.unwrap()?;
+    let notif1: serde_json::Value = serde_json::from_str(notif1.to_text()?)?;
+    let notif2 = ws2.next().await.unwrap()?;
+    let notif2: serde_json::Value = serde_json::from_str(notif2.to_text()?)?;
+
+    assert_eq!(notif1["method"], "base_subscription");
+    assert_eq!(notif2["method"], "base_subscription");
+
+    let block1 = &notif1["params"]["result"];
+    let block2 = &notif2["params"]["result"];
+    assert_eq!(block1["number"], "0x1");
+    assert_eq!(block1["number"], block2["number"]);
+    assert_eq!(block1["hash"], block2["hash"]);
 
     Ok(())
 }
